Limpieza de Archivos innecesarios y creaion de carpetas para ordenar todos los componentes
