+++ src/components/Layout/MainContent.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import '../styles/MainContent.css'
 
 export const MainContent = () => {
