--- src/routes/Live.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Live = () => {
-  return (
-    <div>Live</div>
-  )
-}
