--- src/routes/Podcast.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Podcast = () => {
-  return (
-    <div>Podcast</div>
-  )
-}
