--- src/routes/Radio.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Radio = () => {
-  return (
-    <div>Radio</div>
-  )
-}
